emergency vehicles
so i was only originally going to touch the ambulance and fix how it loads patients because i'm working on a complete forensics and autopsy overhaul, but then my ADHD took over and one thing led to another so i've completely refactored the emergency vehicle code. a word of warning, i didn't really fix anything with the main car.dm so all the bugs that were present there are also present here, it likely would of been easier to fucking redesign this shit but maybe i'll add that to my to-do list later.

- All emergency vehicles have been removed from the main car.dm, and placed under an emergency basetype. This is because car.dm and I have generational beef.
- The lights/sirens system now properly loops and the white halo is generally eliminated through the combination of overlays and lighting. The sirens will automatically cut off if the engine turns off, or the driver exits, yet the lights will remain. You also cannot turn on the sirens unless there's a driver and a siren.
- Police cruisers now have a weapons rack which contains an AR-15 and Shotgun, this can be accessed with ctrl+click when the police cruiser is unlocked.
- The unmarked police cruiser now has random sprites, and does not have a weapons rack.
- The ambulance can now load patients on rollerbeds, and in bodybags, inside the ambulance. The ambulance also has lights and sirens.

i also went insane halfway through and so there are probably some areas that aren't the cleanest. but that's basically it. took way longer than needed trying to fight that motherfucking white halo light